feat: Add bookshelf category selection functionality
Add category selection for bookshelf at main app page. Each time
the category changes the data is updated on the remote server and
in local state book reference, which allows the book to move between
shelves. Since the data from the server is being updated, when the
browser is refreshed the same updated information will be displayed.