feat: Add EntropyInput component
The EntropyInput component is used to generate a unique
address by moving the mouse/finger randomly over its area.
The user movements serve as an extra entropy to the
hash algorithm. I am using the bitcoinjs-lib to create a
bitcoin address. Since the API works with a unique token,
the intention is to use this bitcoin address as the unique token
for storing the client bookshelf data on the backend server.